+++ src/components/10-atoms/textarea/demo.js
@@ -1,9 +1,13 @@
 /* global document */
 import { storiesOf } from '@storybook/html';
+import { boolean, text, withKnobs } from '@storybook/addon-knobs';
+import { html, render } from 'lit-html';
 import Readme from './README.md';
 import './index';
 
+
 storiesOf('Atoms/Textarea/Demos', module)
+  .addDecorator(withKnobs)
   .addParameters({
     readme: {
       sidebar: Readme,
@@ -36,4 +40,40 @@ storiesOf('Atoms/Textarea/Demos', module)
   .add(
     'Textarea - children (default Value)',
     () => `<axa-textarea>prefilled value</axa-textarea>`
-  );
+  )
+  .add('Textarea maxLength added later', () => {
+    const label = text('label*', '');
+    const name = text('name*', '');
+    const refId = text('refId', '');
+    const placeholder = text('placeholder', '');
+    const error = text('error', '');
+    const checkMark = boolean('checkmark', false);
+    const disabled = boolean('disabled', false);
+    const required = boolean('required', false);
+    const invalid = boolean('invalid', false);
+
+    const wrapper = document.createElement('div');
+    const template = html`
+      <axa-textarea
+        refid="${refId}"
+        name="${name}"
+        label="${label}"
+        placeholder="${placeholder}"
+        error="${error}"
+        ?checkmark="${checkMark}"
+        ?disabled="${disabled}"
+        ?required="${required}"
+        ?invalid="${invalid}"
+      ></axa-textarea>
+    `;
+
+    setTimeout(() => {
+
+      document.querySelector('axa-textarea').setAttribute('counter', 'Still ##counter## left');
+      document.querySelector('axa-textarea').setAttribute('maxLength', 100);
+      document.querySelector('axa-textarea').setAttribute('countermax', 'Max maxLength reached');
+    }, 2000);
+
+    render(template, wrapper);
+    return wrapper;
+  });
